fix nan error

+++ dpdl/models/hugging_face_models.py
@@ -75,7 +75,7 @@ def download_generic_huggingface_model(model_name, quantization, trust_remote_co
 
     load_kwargs = {
         "device_map": "auto",
-        "torch_dtype": torch.float16,
+        "torch_dtype": torch.bfloat16,
         "trust_remote_code": trust_remote_code,
     }
